# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ms-notification/config/scheduler_timing": {
            "post": {
                "description": "UpdateScheduledTiming",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UpdateScheduledTiming"
                ],
                "summary": "Update Scheduled Timing config",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "schedulerTiming",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateSchedulerTimingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.SchedulerTiming"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.SchedulerTiming": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "org_id": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateSchedulerTimingRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "org_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        }
    }
}